requirements.txt
Appium-Python-Client==3.2.1
Faker==22.5.0
psycopg2==2.9.9
robotframework==6.1.1
robotframework-appiumlibrary==2.0.0
robotframework-browser==17.5.2
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.6
python_dotenv==1.0.1